refactor - to separate util/path.hpp and util/path.cpp to path/branch.hpp and path/branch.cpp also test/*
